Extract child components for filtering
